refactor: store in dir for mui init content
